requirements.txt
# --- Core Backend ---
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.12.0
pydantic-settings==2.2.1
python-dotenv==1.0.1
requests==2.31.0
httpx==0.28.1
python-jose==3.3.0
bcrypt==4.1.2
tenacity
email-validator==2.3.0
python-multipart==0.0.20

# --- Integrations ---
sendgrid==6.11.0
jinja2==3.1.6
cohere==5.5.8

# --- AI & Vector ---
# Using CPU-only PyTorch to minimize image size (~900MB instead of 3GB)
torch --index-url https://download.pytorch.org/whl/cpu
transformers
huggingface-hub
pgvector
langdetect
numpy<2

# --- Ops ---
prometheus-fastapi-instrumentator==6.1.0
watchfiles==1.1.0